update from dict and exceptions
