+++ erniebot-agent/erniebot_agent/tools/baizhong_tool.py
@@ -1,11 +1,10 @@
 from __future__ import annotations
 
-from typing import Any, Dict, List, Optional, Type
+from typing import Any, List, Optional, Type
 
 from erniebot_agent.messages import AIMessage, HumanMessage
 from erniebot_agent.tools.schema import ToolParameterView
-from pydantic import Field, create_model
-from pydantic.fields import FieldInfo
+from pydantic import Field
 
 from .base import Tool
 
@@ -14,45 +13,15 @@ class BaizhongSearchToolInputView(ToolParameterView):
     query: str = Field(description="Query")
     top_k: int = Field(description="Number of results to return")
 
-    @classmethod
-    def from_dict(cls, field_map: Dict[str, Any]):
-        fields = {}
-        for field_name, field_dict in field_map.items():
-            field_type = field_dict["type"]
-            description = field_dict["description"]
-            field = FieldInfo(annotation=field_type, description=description)
-            fields[field_name] = (field_type, field)
-        return create_model(cls.__name__, __base__=ToolParameterView, **fields)
-
 
 class SearchResponseDocument(ToolParameterView):
     id: str = Field(description="text id")
     title: str = Field(description="title")
     document: str = Field(description="content")
 
-    @classmethod
-    def from_dict(cls, field_map: Dict[str, Any]):
-        fields = {}
-        for field_name, field_dict in field_map.items():
-            field_type = field_dict["type"]
-            description = field_dict["description"]
-            field = FieldInfo(annotation=field_type, description=description)
-            fields[field_name] = (field_type, field)
-        return create_model(cls.__name__, __base__=ToolParameterView, **fields)
-
 
 class BaizhongSearchToolOutputView(ToolParameterView):
     documents: List[SearchResponseDocument] = Field(description="research results")
-
-    @classmethod
-    def from_dict(cls, field_map: Dict[str, Any]):
-        fields = {}
-        for field_name, field_dict in field_map.items():
-            field_type = field_dict["type"]
-            description = field_dict["description"]
-            field = FieldInfo(annotation=field_type, description=description)
-            fields[field_name] = (field_type, field)
-        return create_model(cls.__name__, __base__=ToolParameterView, **fields)
 
 
 class BaizhongSearchTool(Tool):
