+++ erniebot-agent/erniebot_agent/tools/schema.py
@@ -16,7 +16,7 @@ from __future__ import annotations
 
 import inspect
 from dataclasses import dataclass
-from typing import List, Optional, Type, get_args
+from typing import Any, Dict, List, Optional, Type, get_args
 
 from pydantic import BaseModel, Field, create_model
 from pydantic.fields import FieldInfo
@@ -376,6 +376,16 @@ class RemoteToolView:
         if self.returns is not None:
             inputs["responses"] = self.returns.function_call_schema()  # type: ignore
         return scrub_dict(inputs) or {}
+
+    @classmethod
+    def from_dict(cls, field_map: Dict[str, Any]):
+        fields = {}
+        for field_name, field_dict in field_map.items():
+            field_type = field_dict["type"]
+            description = field_dict["description"]
+            field = FieldInfo(annotation=field_type, description=description)
+            fields[field_name] = (field_type, field)
+        return create_model(cls.__name__, __base__=ToolParameterView, **fields)
 
 
 @dataclass
